atualizacoes em classes, e construcaos dos DTOS
