+++ src/main/java/org/serratec/backend/dto/ItemDTO.java
@@ -1,0 +1,4 @@
+package org.serratec.backend.dto;
+
+public class ItemDTO {
+}
