+++ src/main/java/org/serratec/backend/entity/pk/PedidoProdutopk.java
@@ -1,0 +1,4 @@
+package org.serratec.backend.entity.pk;
+
+public class PedidoProdutopk {
+}
